Support for adding custom early "stop" layers
By default (previous versions), only layers including and downstream of STOP_LAYERS were considered to be in "stop_segment_idxs" and involved full computations, i.e. equivalent to a naive implementation. In some cases however, the perturbed region can span a large fraction of the sequence length much earlier (e.g. BPNet with dilated convs) and so it would be helpful to terminate the fastISM mechanism upstream of a STOP_LAYER, and instead at a conv layer beyond which fastISM can not offer much speedup (and likely slows down due to overheads). For this, a mechanism to declare layers, the segments including and downstream of which would be added to "stop_segment_idxs" and would be computed like in a naive implementation.

As of now there is no way to automatically detect when to incorporate an early stop layer (based on fraction of sequence perturbed), but rather manual specification of a reasonable early stop layer.

To achieve this:
- added a way to input names of layer/s to be treated as early stop layers.
- After initial assignment of segment idxs as before, all segments including and downstream layers in early stop layers (if any) are added to stop_segment_idxs
- When change ranges are computed, for a given segment, only inbound segments that are not in stop_segment_idxs are used. This is because once a stop segment is encountered, it is not processed further since no slice assign layers would be required when connecting that segment with other stop segments.
- Intermdediate output models will not return outputs at the junction of two stop segments
- for fast ISM model, Conv1D padding is not set to valid if the conv layer is within a stop segment. Cropping1D is added if within a stop segment.
- SliceAssign layers are not added at the junction of two stop segments
- Many function definitions updated to add in early_stop_segments and stop_segment_idxs
- Left more comments/description line 928 of fast_ism_utils.py

It works with the "full" BPNet model (Anna's architecture) which involves dense net like skip connections. Have not added to test cases, need to do that.