Refined previous commit condition
It is not always possible to have see through layers followed by
conv in same block. Specifically, if conv padding is not valid,
then it would need extra padding which can only be done in beggining
of layer. In fact without this change, an assertion is broken for the
Enformer model for a segment that consists of maxpool -> conv.